add fav/common bases hook and add to comp
